% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/teamlucc-package.R
\docType{data}
\name{ASTER_V002_EAST}
\alias{ASTER_V002_EAST}
\title{Subset of ASTER Digital Elevation Model V002}
\description{
Subset of ASTER Digital Elevation Model V002
}
\seealso{
ASTER_V002_WEST
}

