% Generated by roxygen2 (4.0.1): do not edit by hand
\docType{data}
\name{ASTER_V002_WEST}
\alias{ASTER_V002_WEST}
\title{Subset of ASTER Digital Elevation Model V002}
\description{
Subset of ASTER Digital Elevation Model V002

Subset of ASTER Digital Elevation Model V002
}
\seealso{
ASTER_V002_EAST

ASTER_V002_EAST
}

