% Generated by roxygen2 (4.0.0): do not edit by hand
\name{auto_chg_detect}
\alias{auto_chg_detect}
\title{Perform change detection for two Landsat CDR surface reflectance images}
\usage{
auto_chg_detect(t1_classes, t1_probs, t2_probs, output_basename, output_path,
  n_cpus = 1, overwrite = FALSE, notify = print)
}
\arguments{
\item{t1_classes}{cover classes as output from \code{auto_classify_image}
for time 1 image}

\item{t1_probs}{per class probabilities as output from
\code{auto_classify_image} for time 1 image}

\item{t2_probs}{per class probabilities as output from
\code{auto_classify_image} for time 2 image}

\item{output_path}{the path to use for the output}

\item{output_basename}{the base filename for output files from
\code{auto_chg_detect} (without an extension)}

\item{n_cpus}{the number of CPUs to use for processes that can run in
parallel}

\item{overwrite}{whether to overwrite existing files (otherwise an error
will be raised)}

\item{notify}{notifier to use (defaults to \code{print} function). See the
\code{notifyR} package for one way of sending notifications from R. The
\code{notify} function should accept a string as the only argument.}
}
\description{
First the images should be classified using the \code{auto_classify}
function.
}

