% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/auto_gap_fill.R
\name{auto_gap_fill}
\alias{auto_gap_fill}
\title{Automated removal of gaps in SLC-off images using GNSPI}
\usage{
auto_gap_fill(data_dir, wrspath, wrsrow, start_date, end_date,
  base_date = NULL, tc = TRUE, threshold = 1, n_cpus = 1,
  notify = print, verbose = TRUE, ...)
}
\arguments{
\item{data_dir}{folder where input images are located, with filenames as
output by the \code{\link{auto_preprocess_landsat}} function. This folder
will be searched recursively for images (taking the below path/row, date,
and topographic correction options into account).}

\item{wrspath}{World Reference System (WRS) path}

\item{wrsrow}{World Reference System (WRS) row}

\item{start_date}{start date of period from which images will be chosen to
fill cloudy areas in the base image (as \code{Date} object)}

\item{end_date}{end date of period from which images will be chosen to fill
cloudy areas in the the base image (as \code{Date} object)}

\item{base_date}{ideal date for base image (base image will be chosen as the
image among the available images that is closest to this date). If NULL,
then the base image will be the image with the lowest cloud cover.}

\item{tc}{if \code{TRUE}, use topographically corrected imagery as output by
\code{auto_preprocess_landsat}. IF \code{FALSE} use bands 1-5 and 7 surface
reflectance as output by \code{unstack_ledaps} or
\code{auto_preprocess_landsat} (if \code{auto_preprocess_landsat} was also
run with tc=FALSE).}

\item{threshold}{maximum percent gap allowable in base image. Gap fill will
not occur unless percent gap in base image is greater than this value.}

\item{n_cpus}{the number of CPUs to use for processes that can run in
parallel}

\item{notify}{notifier to use (defaults to \code{print} function).  See the
\code{notifyR} package for one way of sending notifications from R.  The
\code{notify} function should accept a string as the only argument.}

\item{verbose}{whether to print detailed status messages}

\item{...}{additional arguments passed to \code{\link{fill_gaps}}, such as
\code{DN_min}, \code{DN_max}, \code{use_IDL}, \code{verbose}, etc. See
\code{\link{fill_gaps}}.}
}
\value{
\code{Raster*} object with gap filled image.
}
\description{
Uses the GNSPI algorithm from Zhu et al. See \code{\link{fill_gaps}} for
details.  In hilly areas, gap fill should be done after topographic
correction.
}
\details{
The \code{auto_gap_fill} function allows an analyst to automatically
construct a gap-filled image after specifying: \code{data_dir} (a folder of
Landsat images), \code{wrspath} and \code{wrsrow} (the WRS-2 path/row to
use), and \code{start_date} and \code{end_date} (a start and end date
limiting the images to use in the algorithm).  The analyst can also
optionally specify a \code{base_date}, and the \code{auto_gap_fill} function
will automatically pick the image closest to that date to use as the base
image.

As the \code{auto_gap_fill} function automatically chooses images for
inclusion in the gap fill process, it relies on having images stored on disk
in a particular way. To ensure that images are correctly stored on your hard
disk, use the \code{\link{auto_preprocess_landsat}} function to extract the
original Landsat CDR hdf files from the USGS archive. The
\code{auto_preprocess_landsat} function will ensure that images are
extracted and renamed properly so that they can be used with the
\code{auto_gap_fill} script.
}
\references{
Zhu, X., Liu, D., Chen, J., 2012. A new geostatistical approach
for filling gaps in Landsat ETM+ SLC-off images. Remote Sensing of
Environment 124, 49--60.
}

