% Generated by roxygen2 (4.0.1): do not edit by hand
\name{auto_setup_dem}
\alias{auto_setup_dem}
\title{Setup the DEM mosaic for a particular TEAM site}
\usage{
auto_setup_dem(aoi, output_path, dem_extents, n_cpus = 1, overwrite = FALSE,
  crop_to_aoi = FALSE, notify = print, verbose = FALSE)
}
\arguments{
\item{aoi}{area of interest (AOI), as a \code{SpatialPolygonsDataFrame}, to
use as as bounding box when selecting DEMs. Also used to crop and set
projection of the output DEM(s) if \code{crop_to_aoi=TRUE}.}

\item{output_path}{the path to use for the output}

\item{dem_extents}{a \code{SpatialPolygonsDataFrame} of the extents and
filenames for a set of locally available DEM raster(s) that cover the
\code{aoi}. See the \code{\link{get_extent_polys}} function for one means of
generating this list. \code{dem_extents} must have a "filename" column.}

\item{n_cpus}{the number of CPUs to use for processes that can run in
parallel}

\item{overwrite}{whether to overwrite existing files (otherwise an error
will be raised)}

\item{crop_to_aoi}{whether to crop the dem to the supplied AOI, or to the
Landsat path/row polygon for that particular path/row}

\item{notify}{notifier to use (defaults to \code{print} function). See the
\code{notifyR} package for one way of sending notifications from R. The
\code{notify} function should accept a string as the only argument.}

\item{verbose}{whether to print detailed status messages and timing
information}
}
\description{
Setup the DEM mosaic for a particular TEAM site
}

