% Generated by roxygen2 (4.0.1): do not edit by hand
\name{browse_image}
\alias{browse_image}
\title{Simple function to output browse image as PNG}
\usage{
browse_image(x, filename, m = NULL, maxpixels = 5e+05, DN_min = 0,
  DN_max = 255, r = 3, g = 2, b = 1, width = 1200, height = 1200,
  x_fun = NULL, m_fun = NULL)
}
\arguments{
\item{x}{input \code{RasterBrick} or \code{RasterStack} with at least three
bands}

\item{filename}{a filename ending in .png}

\item{m}{an optional mask \code{RasterLayer} to output below the browse
image}

\item{maxpixels}{maximum number of pixels to use in plotting}

\item{DN_min}{minimum DN value}

\item{DN_max}{maximum DN value}

\item{r}{index in \code{x} of the band to use as the red band}

\item{g}{index in \code{x} of the band to use as the green band}

\item{b}{index in \code{x} of the band to use as the blue band}

\item{width}{width in pixels of output PNG file}

\item{height}{height in pixels of output PNG file}

\item{x_fun}{an optional function to apply to \code{x} after x is resampled
according to \code{maxpixels}}

\item{m_fun}{an optional function to apply to \code{m} after m is resampled
according to \code{maxpixels}}
}
\value{
nothing - used for side-effect of saving browse image
}
\description{
Simple function to output browse image as PNG
}

