% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{calc_chg_dir}
\alias{calc_chg_dir}
\title{Calculate change direction}
\usage{
calc_chg_dir(t1p, t2p)
}
\arguments{
\item{t1p}{time 1 posterior probability matrix (with pixels in rows, bands
in columns)}

\item{t2p}{time 2 posterior probability matrix (with pixels in rows, bands
in columns)}
}
\value{
vector of change directions
}
\description{
This code calculate the change direction from two probability images. Not
intended to be called directly - see \code{chg_dir}.
}
\references{
Chen, J., P. Gong, C. He, R. Pu, and P. Shi. 2003.
Land-use/land-cover change detection using improved change-vector analysis.
Photogrammetric Engineering and Remote Sensing 69:369-380.

Chen, J., X. Chen, X. Cui, and J. Chen. 2011. Change vector analysis in
posterior probability space: a new method for land cover change detection.
IEEE Geoscience and Remote Sensing Letters 8:317-321.
}

