% Generated by roxygen2 (4.0.1): do not edit by hand
\name{chg_traj}
\alias{chg_traj}
\title{Calculate change-trajectory image}
\usage{
chg_traj(initial, chg_mag, chg_dir, chg_threshold, classnames = NULL,
  ignorepersistence = TRUE, filename = NULL, ...)
}
\arguments{
\item{initial}{initial cover class as \code{RasterLayer}}

\item{chg_mag}{change magnitude \code{RasterLayer} from \code{CVAPS}}

\item{chg_dir}{change direction \code{RasterLayer} from \code{CVAPS}}

\item{chg_threshold}{the threshold to use as a minimum when determining change
areas (can use \code{DFPS} to determine this value).}

\item{classnames}{an optional vector of classnames to output with the
returned trajectory lookup table}

\item{ignorepersistence}{whether to ignore persistence of a class (if
ignored all pixels where a class persists will be set to NA)}

\item{filename}{filename to save the output \code{RasterLayer} to disk
(optional)}

\item{...}{additional parameters to pass to rasterEngine}
}
\value{
a table of all possible trajectories, with their \code{classnames}
(if specified) and the integer codes used to indicate specific trajectories
in the output image.
}
\description{
Calculate change-trajectory image
}
\details{
Processing can be done in parallel using all using the cluster
facilities in the \code{spatial.tools} package. To enable clustering, call
\code{beginCluster} before running \code{classify}.  To stop the
cluster when finished, call \code{endCluster}.
}
\references{
Chen, J., P. Gong, C.  He, R.  Pu, and P.  Shi.  2003.
Land-use/land-cover change detection using improved change-vector analysis.
Photogrammetric Engineering and Remote Sensing 69:369-380.

Chen, J., X. Chen, X. Cui, and J. Chen. 2011. Change vector analysis in
posterior probability space: a new method for land cover change detection.
IEEE Geoscience and Remote Sensing Letters 8:317-321.
}

