% Generated by roxygen2 (4.0.1): do not edit by hand
\name{classify}
\alias{classify}
\title{Classifies an image using a trained random forest classifier}
\usage{
classify(x, model, classes_file, prob_file, overwrite = FALSE)
}
\arguments{
\item{x}{a \code{Raster*} image with the predictor layer(s) for the
classification}

\item{model}{a trained random forest model as output by
\code{\link{train_classifier}}}

\item{classes_file}{filename for predicted classes (or missing)}

\item{prob_file}{filename for predicted probabilities (or missing) CURRENTLY
IGNORED}

\item{overwrite}{whether to overwrite \code{out_name} if it already exists}
}
\value{
a list with 2 elements: the predicted classes as a
\code{RasterLayer} and the class probabilities as a \code{RasterBrick}
}
\description{
Classifies an image using a trained random forest classifier
}
\examples{
train_data <- get_pixels(L5TSR_1986, L5TSR_1986_2001_training, "class_1986",
                         training=.6)
model <- train_classifier(train_data)
preds <- classify(L5TSR_1986, model)
plot(preds$classes)
plot(preds$probs)
}

