\name{color_image}
\alias{color_image}
\title{Assign class labels to classification file}
\usage{
color_image(x, cls, outfile)
}
\arguments{
  \item{x}{classified image as \code{RasterLayer}}

  \item{cls}{two column matrix, where the first column is
  the class codes (integers) and the second column is the
  class names}

  \item{outfile}{the filename to use for the output}
}
\description{
Assign class labels to classification file
}
\examples{
#TODO: Add examples
}

