% Generated by roxygen2 (4.0.2): do not edit by hand
\name{compcont}
\alias{compcont}
\title{Calculate a contingency table using the composite operator}
\usage{
compcont()
}
\value{
matrix with contingency table
}
\description{
This function calculates a cross tabulation for a map comparison using the
composite operator recommended by Pontius and Cheuk (2006).
}
\references{
Pontius, R. G., and M. L. Cheuk. 2006.  A generalized
cross-tabulation matrix to compare soft-classified maps at multiple
resolutions. International Journal of Geographical Information Science
20:1-30.
}

