% Generated by roxygen2 (4.0.1): do not edit by hand
\docType{methods}
\name{[,pixel_data,character,ANY,ANY-method}
\alias{[,pixel_data,character,ANY,ANY-method}
\title{Extract part of pixel_data class}
\usage{
\S4method{[}{pixel_data,character,ANY,ANY}(x, i, j, ..., drop = TRUE)
}
\arguments{
\item{x}{a \code{pixel_data} object}

\item{i}{a class or list of classes to extract}

\item{j}{unused}

\item{...}{additional arguments (none implemented)}

\item{drop}{For matrices and arrays.  If \code{TRUE} the result is
    coerced to the lowest possible dimension (see the examples).  This
    only works for extracting elements, not for the replacement.  See
    \code{\link{drop}} for further details.
  }
}
\description{
Extract part of pixel_data class
}

