\name{extract_MODIS_sample}
\alias{extract_MODIS_sample}
\title{Generate training data at Landsat scale from a MODIS GLCT image.}
\usage{
extract_MODIS_sample(x, y)
}
\arguments{
  \item{x}{a map of land cover classes as a
  \code{RasterLayer}}
}
\description{
Generate training data at Landsat scale from a MODIS GLCT
image.
}

