% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/get_band_names_from_hdr.R
\name{get_band_names_from_hdr}
\alias{get_band_names_from_hdr}
\title{Extract the band names listed in an ENVI format header file (.hdr)}
\usage{
get_band_names_from_hdr(hdr_file)
}
\arguments{
\item{hdr_file}{an ENVI format header file with a .hdr extension}
}
\value{
A /code{list} of band names extracted from the /code{hdr_file}
}
\description{
Extract the band names listed in an ENVI format header file (.hdr)
}

