\name{get_extent_poly}
\alias{get_extent_poly}
\title{Makes a polygon SpatialPolygonDataFrame of a \code{Raster*} extent.}
\usage{
get_extent_poly(x, fields = c())
}
\arguments{
  \item{x}{A \code{Raster*} object}

  \item{fields}{list of fields to include in the output
  shapefile}
}
\description{
Makes a polygon SpatialPolygonDataFrame of a \code{Raster*}
extent.
}

