% Generated by roxygen2 (4.0.1): do not edit by hand
\name{get_pixels}
\alias{get_pixels}
\title{Extract observed data for use in a classification (training or testing)}
\usage{
get_pixels(x, polys, class_col, training = 1, src = "none")
}
\arguments{
\item{x}{a \code{Raster*} object from which observed data will be extracted.
The data will be extracted from each layer in a \code{RasterBrick} or
\code{RasterStack}.}

\item{polys}{a \code{SpatialPolygonsDataFrame} with polygons, each of which
has been assigned to a particular class (using the \code{class_col}}

\item{class_col}{the name of the column containing the response variable
(for example the land cover type of each pixel)}

\item{training}{indicator of which polygons to use in training. Can be: 1) a
string giving the name of a column indicating whether each polygon is to be
used in training (column equal to TRUE) or in testing (column equal to
FALSE), or 2) a logical vector of length equal to length(polys), or 3) a
number between 0 and 1 indicating the fraction of the polygons to be
randomly selected for use in training.}

\item{src}{name of this data source. Useful when gathering training
data from multiple images.}
}
\value{
data.frame with the training data. Each row will contain the
response (the column chosen by \code{class_col}) as the first column, with
the remaining columns containing the values at that location of each band in
the raster stack.
}
\description{
Extract observed data for use in a classification (training or testing)
}
\examples{
set.seed(1)
train_data <- get_pixels(L5TSR_1986, L5TSR_1986_2001_training, "class_1986",
                         training=.6)
}

