\docType{data}
\name{glcm_test_raster_ENVI_textures}
\alias{glcm_test_raster_ENVI_textures}
\title{GLCM textures calculated in EXELIS ENVI}
\description{
This is the output from running a "co-occurrence measures"
calculation to calculate GLCM textures in EXELIS ENVI from
the \code{glcm_test_raster} included in the \code{teamr}
package. The following settings were used: window size: 3x3
co-occurrence shift: 1 (x), 1 (y) greyscale quantization
levels: 32 textures to compute: mean, variance,
homogeneity, contrast, dissimilarity, entropy, second
moment, correlation
}

