\docType{data}
\name{L5TSR_2001}
\alias{L5TSR_2001}
\title{Landsat 5 Surface Reflectance Image from January 14, 2001 (path 15, row 53)}
\description{
  Portion of Landsat 5 Surface Reflectance image from the
  Landsat Climate Data Record archive. This subset of the
  image includes only bands 1-4.
}
\seealso{
  L5TSR_1986
}
\keyword{datasets}

