\docType{package}
\name{LDPKR-package}
\alias{LDPKR-package}
\alias{LDPKR}
\title{Overview of the LDPKR package}
\description{A set of scripts for processing remote sensing data to support the 
  work of the Tropical Ecology Assessment and Monitoring (TEAM) Network. 
  Intended to be used in conjunction with the set of Python scripts available 
  on \code{\link[https://github.com/azvoleff/LDPK]{github}}.
}
\seealso{\code{\link[raster]{raster}}, \code{\link[landsat]{landsat}}}
\author{
Alex Zvoleff <azvoleff@conservation.org>
}
\keyword{ package }
