% Generated by roxygen2 (4.0.2): do not edit by hand
\docType{methods}
\name{n_test}
\alias{n_test}
\alias{n_test,pixel_data-method}
\title{Get number of testing pixels in a pixel_data object}
\usage{
n_test(x, classes = levels(x@y))

\S4method{n_test}{pixel_data}(x, classes = levels(x@y))
}
\arguments{
\item{x}{a \code{pixel_data} object}

\item{classes}{specifies a subset of classes in \code{x}}
}
\description{
Get number of testing pixels in a pixel_data object
}

