% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/pixel_data.R
\docType{methods}
\name{n_train}
\alias{n_train}
\alias{n_train,pixel_data-method}
\title{Get number of training pixels in a pixel_data object}
\usage{
n_train(x, classes = levels(x@y))

\S4method{n_train}{pixel_data}(x, classes = levels(x@y))
}
\arguments{
\item{x}{a \code{pixel_data} object}

\item{classes}{specifies a subset of classes in \code{x}}
}
\description{
Get number of training pixels in a pixel_data object
}

