% Generated by roxygen2 (4.0.2): do not edit by hand
\name{normalize}
\alias{normalize}
\title{Normalizes two rasters}
\usage{
normalize(x, y, msk, method = "MA", size = ncell(x))
}
\arguments{
\item{x}{a \code{Raster*} to use as the base image}

\item{y}{a \code{Raster*} to normalize to the base image}

\item{msk}{a \code{RasterLayer} with missing values in \code{x} or in {y}
coded as 1, and all other values coded as 0 (optional)}

\item{method}{the regression method to use (must be a method recognized by
\code{lmodel2}}

\item{size}{the number of pixels to use in developing the model}
}
\value{
a \code{Raster*} of \code{y} normalized to \code{x}
}
\description{
Performs relative normalization on two rasters using model II regression.
Based on the approach in the \code{relnorm} function in the \code{landsat}
package.
}
\details{
This function will run in parallel if a parallel backend is registered with
\code{\link{foreach}}.
}
\examples{
L5TSR_2001_normed_1 <- normalize(L5TSR_1986, L5TSR_2001)
plotRGB(L5TSR_2001_normed_1, stretch='lin')

# Use only half as many pixels to calculate the models
L5TSR_2001_normed_2 <- normalize(L5TSR_1986, L5TSR_2001,
                                 size=ncell(L5TSR_1986)/2)
plotRGB(L5TSR_2001_normed_2, stretch='lin')
}
\references{
Sarah Goslee. Analyzing Remote Sensing Data in {R}: The {landsat} Package.
Journal of Statistical Software, 2011, 43:4, pg 1--25.
http://www.jstatsoft.org/v43/i04/
}

