% Generated by roxygen2 (4.0.0): do not edit by hand
\name{nverts}
\alias{nverts}
\title{Count number of vertices in an sp polygon object}
\usage{
nverts(poly_obj)
}
\arguments{
\item{poly_obj}{an sp polygon object}
}
\value{
the number of vertices in the polygon
}
\description{
Count number of vertices in an sp polygon object
}
\examples{
# TODO: Write an example.
}

