% Generated by roxygen2 (4.0.1): do not edit by hand
\docType{class}
\name{pixel_data-class}
\alias{[,pixel_data,character,ANY,ANY-method}
\alias{length.pixel_data}
\alias{levels.pixel_data}
\alias{pixel_data-class}
\alias{print.pixel_data}
\alias{rbind.pixel_data}
\alias{summary.pixel_data}
\title{A class for representing training or testing data}
\usage{
summary.pixel_data(object, ...)

\method{length}{pixel_data}(x)

levels.pixel_data(x)

print.pixel_data(x, ...)

rbind.pixel_data(x, ...)

\S4method{[}{pixel_data,character,ANY,ANY}(x, i, j, ..., drop = TRUE)
}
\arguments{
\item{object}{a \code{pixel_data} object}

\item{...}{additional arguments (none supported)}

\item{x}{
    object from which to extract element(s) or in which to replace element(s).
  }

\item{i}{
    indices specifying elements to extract or replace.  Indices are
    \code{numeric} or \code{character} vectors or empty (missing) or
    \code{NULL}.  Numeric values are coerced to integer as by
    \code{\link{as.integer}} (and hence truncated towards zero).
    Character vectors will be matched to the \code{\link{names}} of the
    object (or for matrices/arrays, the \code{\link{dimnames}}):
    see \sQuote{Character indices} below for further details.

    For \code{[}-indexing only: \code{i}, \code{j}, \code{\dots} can be
    logical vectors, indicating elements/slices to select.  Such vectors
    are recycled if necessary to match the corresponding extent.
    \code{i}, \code{j}, \code{\dots} can also be negative integers,
    indicating elements/slices to leave out of the selection.

    When indexing arrays by \code{[} a single argument \code{i} can be a
    matrix with as many columns as there are dimensions of \code{x}; the
    result is then a vector with elements corresponding to the sets of
    indices in each row of \code{i}.

    An index value of \code{NULL} is treated as if it were \code{integer(0)}.
  }

\item{drop}{For matrices and arrays.  If \code{TRUE} the result is
    coerced to the lowest possible dimension (see the examples).  This
    only works for extracting elements, not for the replacement.  See
    \code{\link{drop}} for further details.
  }
}
\description{
Used to represent training data for a machine learning classifier for image
classificaion, or testing data used for testing a classification.
}
\section{Slots}{

\describe{
\item{\code{x}}{a \code{data.frame} of independent variables (usually pixel values)}

\item{\code{y}}{a \code{data.frame} of the dependent variable (usually land cover
classes)}

\item{\code{pixel_src}}{a data.framevector used to link pixels in \code{x} and
\code{y} to an input polygon}

\item{\code{training_flag}}{a binary vector of length equal to \code{nrow(x)}
indicating each row in x should be used in training (TRUE) or in testing
(FALSE)}

\item{\code{polys}}{a \code{SpatialPolygonsDataFrame} of the polygons used to choose
the pixels in \code{x} and \code{y}.}
}}

