\name{plot_ee}
\alias{plot_ee}
\title{Plot EarthExplorer metadata file}
\usage{
plot_ee(x, start_year, end_year, min_clear = 0.7)
}
\arguments{
  \item{eedf}{a list of Landsat scenes as output from the
  save metadata function on http://earthexplorer.usgs.gov,
  as a \code{data.frame}}
}
\value{
used for side effect of producing a plot
}
\description{
Plot EarthExplorer metadata file
}

