% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/track_time.R
\name{print.Track_time}
\alias{print.Track_time}
\title{Print a Track_time object}
\usage{
\method{print}{Track_time}(x, label, ...)
}
\arguments{
\item{x}{a Track_time object}

\item{label}{(optional) selects a specific tracking timer to print}

\item{...}{ignored}
}
\description{
Print a Track_time object
}

