% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/chg_traj.R
\name{qgis_colormap}
\alias{qgis_colormap}
\title{Calculate Quantum GIS format color map file}
\usage{
qgis_colormap(x, out_file, color_codes, minlength = 5)
}
\arguments{
\item{x}{a lookup table as output by \code{traj_lut}}

\item{out_file}{filename for the output file}

\item{color_codes}{an (optional) list of colors to use in the color table. Should 
have the same length as the number of rows in x.}

\item{minlength}{number of characters used to abbreviate t0_name and t1_name 
columns when building transition names (only used it t0_name and t1_name are 
present in \code{x})}
}
\description{
This function will format a lookup table (lut) into a QGIS color map 
suitable for import into Quantum GIS (QGIS) for color coding a raster file 
using the "singleband pseudocolor" band rendering option.
}
\examples{
lut <- traj_lut(c(1, 2), c("NonForest", "Forest"))
}

