% Generated by roxygen2 (4.0.1): do not edit by hand
\name{rf_classify}
\alias{rf_classify}
\title{Classifies an image using a trained random forest classifier}
\usage{
rf_classify(x, model, class_probs = TRUE, notify = print)
}
\arguments{
\item{x}{a \code{Raster*} image with the predictor layer(s) for the
classification}

\item{model}{a trained random forest model as output by
\code{\link{rf_train}}}

\item{class_probs}{whether to also calculate and return the probabilities of
membership for each class}

\item{use_training_flag}{indicates whether to exclude data flagged as
testing data when training the classifier. For this to work the input
train_data \code{data.frame} must have a column named 'training_flag' that
indicates, for each pixel, whether that pixel is a training pixel (coded as
TRUE) or testing pixel (coded as FALSE).}

\item{notify}{notifier to use (defaults to \code{print} function). See the
\code{notifyR} package for one way of sending notifications from R. The
\code{notify} function should accept a string as the only argument.}
}
\value{
a list with 2 elements: the predicted classes as a
\code{RasterLayer} and the class probabilities as a \code{RasterBrick}
}
\description{
Classifies an image using a trained random forest classifier
}
\examples{
#TODO: Add example.
}

