% Generated by roxygen2 (4.0.1): do not edit by hand
\name{rf_train}
\alias{rf_train}
\title{Trains a random forest to classifier}
\usage{
rf_train(x, train_data, class_probs = TRUE, use_training_flag = TRUE,
  train_control = NULL, tune_grid = NULL, use_rfe = FALSE)
}
\arguments{
\item{x}{a \code{Raster*} image with the predictor layer(s) for the
classification}

\item{train_data}{a \code{link{pixel_data}} object}

\item{class_probs}{whether to also calculate and return the probabilities of
membership for each class}

\item{use_training_flag}{indicates whether to exclude data flagged as
testing data when training the classifier. For this to work the input
train_data \code{data.frame} must have a column named 'training_flag' that
indicates, for each pixel, whether that pixel is a training pixel (coded as
TRUE) or testing pixel (coded as FALSE).}

\item{train_control}{default is NULL (reasonable value will be set
automatically).  For details see \code{\link{trainControl}}.}

\item{tune_grid}{the training grid to be used for training the classifier.
Must be a \code{data.frame} with two columns: ".sigma" and ".C".}

\item{use_rfe}{whether to use Recursive Feature Extraction (RFE) as
implemented in the \code{caret} package to select a subset of the input
features to be used in the classification}

\item{notify}{notifier to use (defaults to \code{print} function). See the
\code{notifyR} package for one way of sending notifications from R. The
\code{notify} function should accept a string as the only argument.}
}
\value{
a trained random forest model (as a \code{train} object from the
\code{caret} package)
}
\description{
Trains a random forest to classifier
}

