\name{split_class}
\alias{split_class}
\title{Calculate change-trajectory image}
\usage{
split_class(train_data)
}
\arguments{
  \item{train_data}{a data table with a column labeled 'y'
  with the observed classes, and one or more columns with
  the values of predictor(s) at each location.}
}
\description{
Calculate change-trajectory image
}

