% Generated by roxygen2 (4.0.1): do not edit by hand
\name{src_name<-}
\alias{src_name<-}
\title{Set src_name for a pixel_data object}
\usage{
src_name(x) <- value
}
\arguments{
\item{x}{a \code{pixel_data} object}

\item{src_name}{a new \code{src_name} to assign for pixels in \code{x}}
}
\description{
Set src_name for a pixel_data object
}

