% Generated by roxygen2 (4.0.1): do not edit by hand
\docType{methods}
\name{src_name}
\alias{src_name}
\alias{src_name,pixel_data-method}
\title{Get src_name for a pixel_data object}
\usage{
src_name(x)

\S4method{src_name}{pixel_data}(x)
}
\arguments{
\item{x}{a \code{pixel_data} object
this pixel_data object}
}
\description{
Get src_name for a pixel_data object
}

