% Generated by roxygen2 (4.0.2): do not edit by hand
\docType{methods}
\name{start_timer}
\alias{start_timer}
\alias{start_timer,Track_time,ANY-method}
\alias{start_timer,Track_time,character-method}
\alias{start_timer,Track_time-method}
\title{Start a tracking timer}
\usage{
start_timer(x, label)

\S4method{start_timer}{Track_time,ANY}(x)

\S4method{start_timer}{Track_time,character}(x, label)
}
\arguments{
\item{x}{a \code{Track_time} object}

\item{label}{an optional label used to maintain multiple tracking timers}
}
\value{
Track_time object
}
\description{
The \code{label} is optional. If not supplied the default timer (labelled
"Default") will be used.
}
\examples{
timer <- Track_time()
print(timer)

timer <- start_timer(timer, 'test')

print(timer, 'test')

timer <- stop_timer(timer, 'test')

print(timer)
}
\seealso{
\code{\link{stop_timer}}
}

