% Generated by roxygen2 (4.0.1): do not edit by hand
\docType{methods}
\name{subsample}
\alias{subsample}
\alias{subsample,pixel_data,numeric-method}
\alias{subsample,pixel_data-method}
\title{Subsample a pixel_data object}
\usage{
subsample(x, frac, by_src = TRUE, flag = TRUE, type = "training")

\S4method{subsample}{pixel_data,numeric}(x, frac, by_src = TRUE,
  flag = TRUE, type = "training")
}
\arguments{
\item{x}{a \code{pixel_data} object}

\item{frac}{fraction of pixels to remove}

\item{by_src}{whether to draw samples from within individual source polygons
(\code{by_src=TRUE}) or from within classes alone \code{by_src=FALSE})}

\item{flag}{whether to flag removed data as testing data (\code{flag=TRUE})
or remove it from dataset entirely (\code{flag=FALSE})}

\item{type}{whether to subsample training data (\code{type='training'}) or
testing data (\code{type='testing'}). Whichever type is chosen, the other
type will be left untouched (for example, if \code{type='testing'}, the
training data will not be changed.}
}
\description{
Subsample a pixel_data object
}

