\name{team_fill_clouds}
\alias{team_fill_clouds}
\title{Perform heavy cloud filling}
\usage{
team_fill_clouds()
}
\description{
Calls CLOUD_REMOVE.pro IDL script by Zhu Xiaolin to fill
heavy clouds in Landsat image.
}
\references{
Zhu, X., Gao, F., Liu, D., Chen, J., 2012. A modified
neighborhood similar pixel interpolator approach for
removing thick clouds in Landsat images. Geoscience and
Remote Sensing Letters, IEEE 9, 521-525.
}

