% Generated by roxygen2 (4.0.1): do not edit by hand
\docType{package}
\name{teamlucc-package}
\alias{teamlucc}
\alias{teamlucc-package}
\title{TEAM Remote Sensing Processing Tools}
\description{
TEAM land use and cover change data processing toolkit
}
\details{
teamlucc is a set of routines to support analyzing land use and cover change
(LUCC) in R. The package was designed to support analyzing LUCC in the Zone
of Interaction (ZOIs) of monitoring sites in the Tropical Ecology Assessment
and Monitoring (TEAM) Network.
}
\author{
Alex Zvoleff, \email{azvoleff@conservation.org}
}
\keyword{package}

