\docType{package}
\name{teamr-package}
\alias{teamr}
\alias{teamr-package}
\title{TEAM Remote Sensing Processing Tools}
\description{
TEAM Remote Sensing Processing Tools
}
\details{
A set of scripts for processing remote sensing data to
support the work of the Tropical Ecology Assessment and
Monitoring (TEAM) Network. Intended to be used in
conjunction with the set of Python scripts available in the
\href{https://github.com/azvoleff/teampy}{teampy
repository} on github.
}
\author{
Alex Zvoleff, \email{azvoleff@conservation.org}
}
\keyword{package}

