% Generated by roxygen2 (4.0.1): do not edit by hand
\name{threshold}
\alias{threshold}
\title{Automatically determine value for image thresholding}
\usage{
threshold(x, method = "huang", by = 1, maxpixels = 1e+05)
}
\arguments{
\item{x}{the input image, as a matrix or raster}

\item{method}{the thresholding method. Currently only "huang" is
implemented.}

\item{by}{step size to use when calculating histogram}

\item{maxpixels}{maximum number of pixels size to use when calculating
histogram}
}
\value{
integer threshold value
}
\description{
The only method currently implemented is Huang's fuzzy thresholding method.
The code for Huang's method was ported to C++ for teamluc from the code in
the Auto_threshold imageJ plugin by Gabriel Landini. See original code at:
http:#www.mecourse.com/landinig/software/autothreshold/autothreshold.html
}
\details{
This function will run in parallel if a parallel backend is registered with
\code{\link{foreach}}.
}
\references{
Huang, L.-K., and M.-J. J. Wang. 1995. Image thresholding by
minimizing the measures of fuzziness. Pattern recognition 28 (1):41--51.
}

