% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{threshold_Huang}
\alias{threshold_Huang}
\title{Threshold an image using Huang's fuzzy thresholding method.}
\usage{
threshold_Huang(data)
}
\arguments{
\item{data}{the input image}
}
\value{
integer threshold value
}
\description{
Implements Huang's fuzzy thresholding method. This function is called by
the \code{\link{threshold}} function. It is not intended to be used
directly.
}
\details{
Ported to C++ by from the code in the Auto_threshold imageJ plugin by
Gabriel Landini.

See original code at:
http://www.mecourse.com/landinig/software/autothreshold/autothreshold.html
}
\references{
Huang, L.-K., and M.-J. J. Wang. 1995. Image thresholding by
minimizing the measures of fuzziness. Pattern recognition 28 (1):41--51.
}

