\name{trackTime}
\alias{trackTime}
\title{Print status messages tracking run time for an R script}
\usage{
trackTime(label = "", action = "print", email = FALSE)
}
\arguments{
  \item{an}{optional label used to maintain multiple
  tracking timers}

  \item{action}{a string indicating whether to \code{start}
  the timer or \code{print} (default) the time that has
  passed since the timer was last started.}
}
\description{
Print status messages tracking run time for an R script
}
\examples{
#TODO: Add examples
}

