% Generated by roxygen2 (4.0.1): do not edit by hand
\name{train_classifier}
\alias{train_classifier}
\title{Trains a classifier}
\usage{
train_classifier(train_data, type = "rf", use_training_flag = TRUE,
  train_control = NULL, tune_grid = NULL, use_rfe = FALSE)
}
\arguments{
\item{train_data}{a \code{link{pixel_data}} object}

\item{type}{either "svm" (to fit a support vector machine) or "rf" (to fit a
random forest).}

\item{use_training_flag}{indicates whether to exclude data flagged as
testing data when training the classifier. For this to work the input
train_data \code{data.frame} must have a column named 'training_flag' that
indicates, for each pixel, whether that pixel is a training pixel (coded as
TRUE) or testing pixel (coded as FALSE).}

\item{train_control}{default is NULL (reasonable values will be set
automatically).  For details see \code{\link{trainControl}}.}

\item{tune_grid}{the training grid to be used for training the classifier.
See Details.}

\item{use_rfe}{whether to use Recursive Feature Extraction (RFE) as
implemented in the \code{caret} package to select a subset of the input
features to be used in the classification. NOT YET SUPPORTED.}

\item{notify}{notifier to use (defaults to \code{print} function). See the
\code{notifyR} package for one way of sending notifications from R. The
\code{notify} function should accept a string as the only argument.}
}
\value{
a trained random forest model (as a \code{train} object from the
\code{caret} package)
}
\description{
This function trains a Support Vector Machine (SVM) or Random Forest (RF)
classifier for use in an image classification.
}
\details{
For \code{type='svm'}, \code{tunegrid} must be a \code{data.frame} with two
columns: ".sigma" and ".C". For \code{type='rf'}, must be a
\code{data.frame} with one column: '.mtry'.
}
\examples{
train_data <- get_pixels(L5TSR_1986, L5TSR_1986_2001_training, "class_1986",
                         training=.6)
model <- train_classifier(train_data)
}

