% Generated by roxygen2 (4.0.1): do not edit by hand
\docType{methods}
\name{training_flag}
\alias{training_flag}
\alias{training_flag,pixel_data-method}
\alias{training_flag<-}
\alias{training_flag<-,pixel_data-method}
\title{Get or set training_flag for a pixel_data object}
\usage{
training_flag(x)

\S4method{training_flag}{pixel_data}(x)

training_flag(x) <- value

\S4method{training_flag}{pixel_data}(x) <- value
}
\arguments{
\item{x}{a \code{pixel_data} object}

\item{value}{a new \code{training_flag} to assign for pixels in \code{x}}
}
\description{
Get or set training_flag for a pixel_data object
}

