% Generated by roxygen2 (4.0.0): do not edit by hand
\name{ttsdf2raster}
\alias{ttsdf2raster}
\title{Function to convert TIMESAT tts data.frame an R raster.}
\usage{
ttsdf2raster(x, base_image)
}
\arguments{
\item{x}{A TTS data.frame as output by tts2df}

\item{base_image}{A string giving the location of a raster file to use
for georeferencing the output raster. Use one of the original raster files
that was input to TIMESAT.}
}
\value{
A raster object
}
\description{
Function to convert TIMESAT tts data.frame an R raster.
}
\examples{
# TODO: Need to add examples here, and need to include a sample TIMESAT tpa
# file in the package data.
}

