% Generated by roxygen2 (4.0.0): do not edit by hand
\docType{methods}
\name{summary.accuracy}
\alias{accuracy}
\alias{accuracy,RasterLayer,SpatialPolygonsDataFrame,ANY,character,ANY-method}
\alias{accuracy,RasterLayer,SpatialPolygonsDataFrame,ANY,character-method}
\alias{accuracy,RasterLayer,pixel_data,ANY,missing,ANY-method}
\alias{accuracy,RasterLayer,pixel_data,ANY,missing-method}
\alias{accuracy,train,ANY,ANY,missing,ANY-method}
\alias{accuracy,train,ANY,ANY,missing-method}
\alias{print.accuracy}
\alias{print.summary.accuracy}
\alias{summary.accuracy}
\title{Calculate statistics summarizing classification accuracy}
\usage{
summary.accuracy(object, ...)

print.summary.accuracy(x, ...)

print.accuracy(x, ...)

accuracy(x, test_data, pop, class_col, reclass_mat)

\S4method{accuracy}{train,ANY,ANY,missing}(x, test_data, pop, class_col,
  reclass_mat)

\S4method{accuracy}{RasterLayer,pixel_data,ANY,missing}(x, test_data, pop,
  class_col, reclass_mat)

\S4method{accuracy}{RasterLayer,SpatialPolygonsDataFrame,ANY,character}(x,
  test_data, pop, class_col, reclass_mat)
}
\arguments{
\item{x}{either a classification model with a \code{predict} method or a
\code{RasterLayer} (see Details)}

\item{test_data}{a \code{link{pixel_data}} object,
\code{SpatialPolygonsDataFrame}, or NULL (see Details).}

\item{pop}{A \code{RasterLayer}, \code{numeric} of length equal to the
number of clasess, or NULL (see Details).}

\item{class_col}{required if \code{test_data} is a
\code{SpatialPolygonsDataFrame}. Defines the name of the column containing
the observed cover class IDs}

\item{reclass_mat}{a reclassification matrix to be used in the case of a
model fit by \code{classify_image} with the \code{do_split} option selected}
}
\value{
\code{\link{accuracy-class}} instance
}
\description{
Calculates a contingency table and various statistics for use in image
classification accuracy assessment and map comparison. Contingency table
includes user's, producer's, and overall accuracies for an image
classification, and quantity disagreement \code{Q} and allocation
disagreement \code{A}. \code{Q} and \code{A} are calculated based on Pontius
and Millones (2011). Standard errors for 95 percent confidence intervals for
the user's, producer's and overall accuracies are calculated as in Foody and
Stehman (2009) Table 21.3. To avoid bias due to the use of a sample
contingency table, the contingency table will be converted to a population
contingency table if the variable 'pop' is provided. For an accuracy
assessment using testing data from a simple random sample, 'pop' does not
need to be provided (see Details).
}
\details{
\code{x} can be one of:
\enumerate{

\item A prediction model as output from one of the \code{teamlucc}
  \code{classify_image} functions. If \code{x} is a model, and testing data
  is included in the model, \code{pop} and \code{test_data} can both be
  missing, and accuracy will still run (though the output will in this case
  be biased unless the testing data is from a simple random sample). If
  \code{x} is a \code{RasterLayer}, then \code{test_data} must be supplied.

\item A \code{RasterLayer} with a predicted map.
}

\code{test_data} can be one of:
\enumerate{
  \item \code{NULL}. If test_data is \code{NULL}, \code{accuracy} will try to use
        testing data included in \code{x}. This will only work if \code{x}
        is a model of class \code{train} from the \code{caret} package, and
        if the model was run using the one of the \code{teamlucc}
        \code{classify_image} functions.

\item A \code{SpatialPolygonsDataFrame} object, in which case \code{accuracy}
        will extract the predicted classes within each polygon from \code{x}.
        This will only work if \code{x} is a \code{RasterLayer}.

\item A \code{pixel_data} object, in which case \code{accuracy} will use the
        included \code{training_flag} indicator to separate testing and
        training data.
}

\code{pop} can be one of:
\enumerate{
  \item NULL, in which case the sample frequencies will be used as estimates
        of the population frequencies of each class.

\item A list of length equal to the number of classes in the map giving
        the total number of pixels in the population for each class.

\item A predicted cover map from as a \code{RasterLayer}, from which the
        class frequencies will be tabulated and used as the population
        frequencies.
}
}
\examples{
accuracy(L5TSR_1986_classified$model)
}
\references{
Pontius, R. G., and M. Millones. 2011. Death to Kappa: birth of
quantity disagreement and allocation disagreement for accuracy assessment.
International Journal of Remote Sensing 32:4407-4429.

Olofsson, P., G. M. Foody, S. V. Stehman, and C. E. Woodcock.  2013. Making
better use of accuracy data in land change studies: Estimating accuracy and
area and quantifying uncertainty using stratified estimation.  Remote
Sensing of Environment 129:122-131.

Foody, G.M., Stehman, S.V., 2009. Accuracy Assessment, in: Warner, T.A.,
Nellis, M.D., Foody, G.M. (Eds.), The SAGE Handbook of Remote Sensing. SAGE.
}

