% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/teamlucc-package.R
\docType{data}
\name{ASTER_V002_WEST}
\alias{ASTER_V002_WEST}
\title{Subset of ASTER Digital Elevation Model V002}
\description{
Subset of ASTER Digital Elevation Model V002

Subset of ASTER Digital Elevation Model V002
}
\seealso{
ASTER_V002_EAST

ASTER_V002_EAST
}

