% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/auto_chg_detect.R
\name{auto_chg_detect}
\alias{auto_chg_detect}
\title{Perform change detection for two Landsat CDR surface reflectance images}
\usage{
auto_chg_detect(t1_classes, t1_probs, t2_probs, output_path, output_basename,
  ext = "tif", overwrite = FALSE, chg_threshold = NULL, notify = print)
}
\arguments{
\item{t1_classes}{cover classes as output from \code{auto_classify_image} 
for time 1 image}

\item{t1_probs}{per class probabilities as output from 
\code{auto_classify_image} for time 1 image}

\item{t2_probs}{per class probabilities as output from 
\code{auto_classify_image} for time 2 image}

\item{output_path}{the path to use for the output}

\item{output_basename}{the base filename for output files from 
\code{auto_chg_detect} (without an extension)}

\item{ext}{file extension to use when saving output rasters (determines 
output file format).}

\item{overwrite}{whether to overwrite existing files (otherwise an error
will be raised)}

\item{chg_threshold}{the threshold to use determining change and no-change 
areas from the change magnitude image (see \code{\link{chg_mag}}. If 
\code{NULL}, then \code{\link{threshold}} will be used to dermine this 
threshold value automatically. A threshold in the range of .75-1 is 
recommended as a starting point.}

\item{notify}{notifier to use (defaults to \code{print} function).  See the 
\code{notifyR} package for one way of sending notifications from R.  The 
\code{notify} function should accept a string as the only argument.}
}
\value{
nothing - used for the side effect of performing change detection
}
\description{
This image automates the change detection process using the Change Vector 
Analysis in Posterior Probability Space (CVAPS) algorithm. The threshold for 
change/no-change mapping is determined using Huang's algorithm (see 
\code{\link{threshold}} or can be specified manually. First the images 
should be classified using the \code{auto_classify} function (or any other 
classification approach that yields per-pixel probabilities of class 
membership).
}
\references{
Chen, J., X. Chen, X. Cui, and J. Chen. 2011. Change vector 
analysis in posterior probability space: a new method for land cover change 
detection.  IEEE Geoscience and Remote Sensing Letters 8:317-321.
}

