% Generated by roxygen2 (4.0.1): do not edit by hand
\name{auto_cloud_fill}
\alias{auto_cloud_fill}
\title{Automated removal of clouds from Landsat CDR imagery}
\usage{
auto_cloud_fill(data_dir, wrspath, wrsrow, start_date, end_date, out_name,
  base_date = NULL, tc = TRUE, ext = "tif", sensors = c("L4T", "L5T",
  "L7E", "L8C"), img_type = "CDR", threshold = 1, max_iter = 5,
  notify = print, verbose = 1, overwrite = FALSE, ...)
}
\arguments{
\item{data_dir}{folder where input images are located, with filenames as
output by the \code{\link{auto_preprocess_landsat}} function. This folder
will be searched recursively for images (taking the below path/row, date,
and topographic correction options into account).}

\item{wrspath}{World Reference System (WRS) path}

\item{wrsrow}{World Reference System (WRS) row}

\item{start_date}{start date of period from which images will be chosen to
fill cloudy areas in the base image (as \code{Date} object)}

\item{end_date}{end date of period from which images will be chosen to fill
cloudy areas in the the base image (as \code{Date} object)}

\item{base_date}{ideal date for base image (base image will be chosen as the
image among the available images that is closest to this date). If NULL,
then the base image will be the image with the lowest cloud cover.}

\item{out_name}{base filename (without an extension - see \code{ext}
argument) for cloud filled image.  The mask file for the cloud filled image
will be saved with the same name, with the added suffix "_mask".}

\item{tc}{if \code{TRUE}, use topographically corrected imagery as output by
\code{auto_preprocess_landsat}. IF \code{FALSE} use bands 1-5 and 7 surface
reflectance as output by \code{unstack_ledaps} or
\code{auto_preprocess_landsat} (if \code{auto_preprocess_landsat} was also
run with tc=FALSE).}

\item{ext}{file extension to use when searching for input rasters and when
saving output rasters (determines output file format). Should match file
extension of input rasters (and should most likely match the value chosen
for \code{ext} when \code{auto_preprocess_landsat} was run).}

\item{sensors}{choose the sensors to include when selecting images (useful
for excluding images from a particular satellite if desired). Can be any of
"L4T", "L5T", "L7E", and/or "L8C".}

\item{img_type}{type of Landsat imagery to preprocess. Can be "CDR" for
Landsat Climate Data Record (CDR) imagery in HDR format, or "L1T" for
Standard Terrain Correction (Level 1T) imagery. Note that if L1T imagery is
used, fmask must be run locally (see https://code.google.com/p/fmask) prior
to using \code{auto_preprocess_landsat}.}

\item{threshold}{maximum percent cloud cover allowable in base image. Cloud
fill will iterate until percent cloud cover in base image is below this
value, or until \code{max_iter} iterations have been run}

\item{max_iter}{maximum number of times to run cloud fill script}

\item{notify}{notifier to use (defaults to \code{print} function).  See the
\code{notifyR} package for one way of sending notifications from R.  The
\code{notify} function should accept a string as the only argument.}

\item{verbose}{whether to print detailed status messages. Set to FALSE or 0
for no status messages. Set to 1 for basic status messages. Set to 2 for
detailed status messages.}

\item{overwrite}{whether to overwrite \code{out_name} if it already exists}

\item{...}{additional arguments passed to \code{\link{cloud_remove}}, such
as \code{DN_min}, \code{DN_max}, \code{algorithm}, \code{byblock},
\code{verbose}, etc. See \code{\link{cloud_remove}} for details}
}
\value{
a list with two elements: "filled", a \code{Raster*} object with
cloud filled image, and "mask", a \code{RasterLayer} object with the cloud
mask for the cloud filled image.
}
\description{
Uses one of four cloud reomval algorithms (see \code{\link{cloud_remove}})
to remove thick clouds from Landsat imagery. In hilly areas, topographic
correction should be done before cloud fill.
}
\details{
The \code{auto_cloud_fill} function allows an analyst to automatically
construct a cloud-filled image after specifying: \code{data_dir} (a folder
of Landsat images), \code{wrspath} and \code{wrsrow} (the WRS-2 path/row to
use), and \code{start_date} and \code{end_date} (a start and end date
limiting the images to use in the algorithm).  The analyst can also
optionally specify a \code{base_date}, and the \code{auto_cloud_fill}
function will automatically pick the image closest to that date to use as
the base image.

As the \code{auto_cloud_fill} function automatically chooses images for
inclusion in the cloud fill process, it relies on having images stored on
disk in a particular way, and currently only supports cloud fill for Landsat
CDR surface reflectance images. To ensure that images are correctly stored
on your hard disk, use the \code{\link{auto_preprocess_landsat}} function to
extract the original Landsat CDR hdf files from the USGS archive. The
\code{auto_preprocess_landsat} function will ensure that images are
extracted and renamed properly so that they can be used with the
\code{auto_cloud_fill} script.
}
\references{
Zhu, X., Gao, F., Liu, D., Chen, J., 2012. A modified
neighborhood similar pixel interpolator approach for removing thick clouds
in Landsat images.  Geoscience and Remote Sensing Letters, IEEE 9, 521--525.
doi:10.1109/LGRS.2011.2173290
}

