% Generated by roxygen2 (4.0.2): do not edit by hand
\name{auto_normalize}
\alias{auto_normalize}
\title{Normalize a set of preprocessed CDR images to a base image}
\usage{
auto_normalize(image_files, base, overwrite = FALSE)
}
\arguments{
\item{image_files}{list of filenames for images to normalize}

\item{base}{(optional) filename of base image. If not supplied, the base
image will be automatically chosen from the images in \code{image_files}, as
the image with the lowest percent cloud cover.}

\item{overwrite}{whether to overwrite existing files}
}
\value{
nothing - used for side effect of normalizing imagery
}
\description{
This function uses model II regression to perform relative normalization to
match a set of Landsat CDR surface reflectance images. The function assumes
the images were preprocessed using the \code{auto_preprocess_landsat}
function. A base image can be optionally supplied. If a base image is not
supplied, then the function will calculate the percent cloud cover of each
input image, and automatically choose the image with the least cloud cover
as the base image. This function assumes that the images (and image masks)
were preprocessed using the \code{auto_preprocess_landsat} function.
}
\details{
This function will run in parallel if a parallel backend is registered with
\code{\link{foreach}}.
}

