% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/auto_preprocess_landsat.R
\name{auto_preprocess_landsat}
\alias{auto_preprocess_landsat}
\title{Preprocess surface reflectance imagery from the Landsat CDR archive}
\usage{
auto_preprocess_landsat(image_dirs, prefix, tc = FALSE, dem_path = NULL,
  aoi = NULL, output_path = NULL, mask_type = "fmask",
  mask_output = FALSE, n_cpus = 1, cleartmp = FALSE, overwrite = FALSE,
  of = "GTiff", ext = "tif", notify = print, verbose = FALSE)
}
\arguments{
\item{image_dirs}{list of paths to a set of Landsat CDR image files as 
downloaded from ESPA and extracted by \code{\link{espa_extract}}}

\item{prefix}{string to use as a prefix for all filenames}

\item{tc}{whether to topographically correct imagery (if \code{TRUE}, then 
\code{dem_path} must be specified)}

\item{dem_path}{path to a set of DEMs as output by \code{auto_setup_dem} 
(only required if tc=TRUE)}

\item{aoi}{area of interest (AOI), as a \code{SpatialPolygonsDataFrame}.  If 
supplied, this aoi is used to crop and set the projection system of the 
output. Must be in a projected coordinate system.}

\item{output_path}{the path to use for the output (optional - if NULL then 
output images will be saved alongside the input images in the same folder).}

\item{mask_type}{which cloud mask to use to mask clouds when performing 
topographic correction. Can be one of "fmask", "6S", or "both".  See 
Details.  (Ignored if \code{tc=FALSE)}.}

\item{mask_output}{if \code{TRUE}, cloud, cloud shadow, and fill areas 
(SLC-off gaps and areas with no data) will be set to \code{NA} in the 
output. Note this setting affects the final output file only - cloud, cloud 
shadow, and gap areas are masked out of the image during topographic 
correction regardless of the value of \code{mask_output}.}

\item{n_cpus}{the number of CPUs to use for processes that can run in 
parallel}

\item{cleartmp}{whether to clear temp files on each run through the loop}

\item{overwrite}{whether to overwrite existing files (otherwise an error 
will be raised)}

\item{of}{output format to use when saving output rasters. See description 
of \code{of} in \code{\link{gdalwarp}}.}

\item{ext}{file extension to use when saving output rasters (determines 
output file format). Should match file extension for output format chosen by 
\code{of}.}

\item{notify}{notifier to use (defaults to \code{print} function). See the 
\code{notifyR} package for one way of sending notifications from R. The 
\code{notify} function should accept a string as the only argument.}

\item{verbose}{whether to print detailed status messages and timing 
information}
}
\value{
nothing - used for the side effect of preprocessing imagery
}
\description{
This function preprocesses surface reflectance imagery from the Landsat 
Climate Data Record (CDR) archive. \code{auto_preprocess_landsat} can 
reproject CDR tiles to match the projection of a given \code{aoi}, crop the 
tiles to match the \code{aoi} or a common WRS-2 path/row polygon, mask 
missing data and clouds out of the CDR tiles, and perform topographic 
correction.
}
\details{
\code{mask_type} chooses the cloud mask to use if topographic correction is 
performed (\code{tc=TRUE}). The mask can be one of three different options: 
"6S", "fmask", or "combined". Each option uses a different combination of 
cloud mask layers from the CDR product. The "6S" masks out any areas coded 
as fill (fill_QA=255), cloud (cloud_QA=255), cloud shadow
(cloud_shadow_QA=255) or adjacent to cloud (adjacent_cloud_QA=255). The 
"fmask" option masks out any areas coded as fill (fmask=255), cloud 
(fmask=4) or cloud shadow (fmask=2).  The combined option combines the "6S" 
and "fmask" approaches to masks out areas coded as fill, cloud, cloud 
shadow, or adjacent to cloud using either method. Note that "fmask" is the 
only supported option when \code{file_format} is L1T. Further, if  L1T imagery 
is used, fmask must be run locally (see https://code.google.com/p/fmask) 
prior to using \code{auto_preprocess_landsat}.

Prior to running \code{auto_preprocess_landsat}, \code{\link{espa_extract}} 
should be used to extract the original zipfiles supplied by USGS. To perform 
topographic correction with \code{auto_preprocess_landsat}, first run 
\code{\link{auto_setup_dem}} to preprocess a set of DEM tiles. Then run 
\code{auto_preprocess_landsat} with the \code{tc=TRUE} option.

This function will run in parallel if a parallel backend is registered with 
\code{\link{foreach}}.
}
\seealso{
\code{\link{espa_extract}}, \code{\link{unstack_ledapscdr}}, 
\code{\link{auto_setup_dem}}
}

