% Generated by roxygen2 (4.0.1): do not edit by hand
\name{auto_setup_dem}
\alias{auto_setup_dem}
\title{Setup the DEM mosaic for a given AOI}
\usage{
auto_setup_dem(aoi, output_path, dem_extents, of = "GTiff", ext = "tif",
  n_cpus = 1, overwrite = FALSE, crop_to_aoi = FALSE, notify = print,
  verbose = FALSE)
}
\arguments{
\item{aoi}{area of interest (AOI), as a \code{SpatialPolygonsDataFrame}, to
use as as bounding box when selecting DEMs. Also used to crop and set
projection of the output DEM(s) if \code{crop_to_aoi=TRUE}. Must be in a
projected coordinate system.}

\item{output_path}{the path to use for the output}

\item{dem_extents}{a \code{SpatialPolygonsDataFrame} of the extents and
filenames for a set of locally available DEM raster(s) that cover the
\code{aoi}. See the \code{\link{get_extent_polys}} function for one means of
generating this list. \code{dem_extents} must have a "filename" column.}

\item{of}{output format to use when saving output rasters. See description
of \code{of} in \code{\link{gdalwarp}}.}

\item{ext}{file extension to use when saving output rasters (determines
output file format). Should match file extension for output format chosen by
\code{of}.}

\item{n_cpus}{the number of CPUs to use for processes that can run in
parallel}

\item{overwrite}{whether to overwrite existing files (otherwise an error
will be raised)}

\item{crop_to_aoi}{whether to crop the dem to the supplied AOI, or to the
Landsat path/row polygon for that particular path/row}

\item{notify}{notifier to use (defaults to \code{print} function). See the
\code{notifyR} package for one way of sending notifications from R. The
\code{notify} function should accept a string as the only argument.}

\item{verbose}{whether to print detailed status messages and timing
information}
}
\value{
nothing - used for the side effect of setting up DEMs
}
\description{
This function will setup a set of DEM tiles for each the Landsat path/row
needed to cover a given AOI. The tiles can optionally be cropped to cover
only the portion of each path/row that is included in the AOI, or can cover
the full scene for each path/row needed to cover the AOI.
}
\details{
This function uses \code{gdalUtils}, which requires a local GDAL
installation.  See http://trac.osgeo.org/gdal/wiki/DownloadingGdalBinaries
or http://trac.osgeo.org/osgeo4w/ to download the appropriate installer for
your operating system.
}

