% Generated by roxygen2 (4.0.2): do not edit by hand
\name{browse_image}
\alias{browse_image}
\title{Simple function to make small preview plot from large raster image}
\usage{
browse_image(x, m = NULL, maxpixels = 5e+05, DN_min = 0, DN_max = 255,
  r = 3, g = 2, b = 1, x_fun = NULL, m_fun = NULL)
}
\arguments{
\item{x}{input \code{RasterBrick} or \code{RasterStack} with at least three
bands}

\item{m}{an optional mask \code{RasterLayer} to output below the browse
image}

\item{maxpixels}{maximum number of pixels to use in plotting}

\item{DN_min}{minimum DN value}

\item{DN_max}{maximum DN value}

\item{r}{index in \code{x} of the band to use as the red band}

\item{g}{index in \code{x} of the band to use as the green band}

\item{b}{index in \code{x} of the band to use as the blue band}

\item{x_fun}{an optional function to apply to \code{x} after x is resampled
according to \code{maxpixels}}

\item{m_fun}{an optional function to apply to \code{m} after m is resampled
according to \code{maxpixels}}
}
\value{
nothing - used for side-effect of saving browse image
}
\description{
Simple function to make small preview plot from large raster image
}

