% Generated by roxygen2 (4.0.1): do not edit by hand
\name{chg_traj}
\alias{chg_traj}
\title{Calculate change-trajectory image}
\usage{
chg_traj(chg_mag, chg_dir, chg_threshold, filename, overwrite = FALSE, ...)
}
\arguments{
\item{chg_mag}{change magnitude \code{RasterLayer} from \code{CVAPS}}

\item{chg_dir}{change direction \code{RasterLayer} from \code{CVAPS}}

\item{chg_threshold}{the threshold to use as a minimum when determining change
areas (can use \code{DFPS} to determine this value).}

\item{filename}{filename to save the output \code{RasterLayer} to disk
(optional)}

\item{overwrite}{whether to overwrite existing files (otherwise an error
will be raised)}

\item{...}{additional parameters to pass to rasterEngine}
}
\value{
a {RasterLayer} of change trajectories, with change trajectories
coded as in the \code{lut} output by \code{traj_lut}
}
\description{
This function will calculate trajectories of land cover change using the
Change Vector Analysis in Posterior Probability Space (CVAPS) approach of
comparing posterior probabilities of class membership with an automatically
determined threshold. Areas of no change are coded as -1. A lookup table for
the codes output by \code{chg_traj} can be calculated with \code{traj_lut}.
}
\details{
This function will run in parallel if a parallel backend is registered with
\code{\link{foreach}}.
}
\examples{
\dontrun{
t0_train_data <- get_pixels(L5TSR_1986, L5TSR_1986_2001_training, "class_1986",training=.6)
t0_model <- train_classifier(t0_train_data)
t0_preds <- classify(L5TSR_1986, t0_model)
t1_train_data <- get_pixels(L5TSR_2001, L5TSR_1986_2001_training, "class_2001", training=.6)
t1_model <- train_classifier(t1_train_data)
t1_preds <- classify(L5TSR_2001, t1_model)
t0_t1_chgmag <- chg_mag(t0_preds$probs, t1_preds$probs)
t0_t1_chgdir <- chg_dir(t0_preds$probs, t1_preds$probs)

lut <- traj_lut(t0_preds$codes$code, t0_preds$codes$class)
t0_t1_chgtraj <- chg_traj(lut, t0_t1_chgmag, t0_t1_chgdir, .5)

# Change areas are coded following the above lookup-table (lut):
plot(t0_t1_chgtraj)

# No change areas are -1:
plot(t0_t1_chgtraj == -1)
}
}
\references{
Chen, J., P. Gong, C.  He, R.  Pu, and P.  Shi.  2003.
Land-use/land-cover change detection using improved change-vector analysis.
Photogrammetric Engineering and Remote Sensing 69:369-380.

Chen, J., X. Chen, X. Cui, and J. Chen. 2011. Change vector analysis in
posterior probability space: a new method for land cover change detection.
IEEE Geoscience and Remote Sensing Letters 8:317-321.
}

