% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/class_statistics.R
\name{class_statistics}
\alias{class_statistics}
\title{Exports statistics on pixels within each of a set of land cover classes}
\usage{
class_statistics(x, y, class_col)
}
\arguments{
\item{x}{A \code{RasterLayer} from which class statistics will be 
calculated.}

\item{y}{A \code{SpatialPolygonsDataFrame} with cover class 
polygons}

\item{class_col}{the name of the column containing the response variable 
(for example the land cover type of each pixel)}
}
\value{
A data.frame of class statistics.
}
\description{
Exports statistics on pixels within each of a set of land cover classes
}
\examples{
class_statistics(L5TSR_1986, L5TSR_1986_2001_training, "class_1986")
}

