% Generated by roxygen2 (4.0.1): do not edit by hand
\name{classify}
\alias{classify}
\title{Classifies an image using a trained random forest classifier}
\usage{
classify(x, model, classes_file, prob_file, overwrite = FALSE,
  notify = print)
}
\arguments{
\item{x}{a \code{Raster*} image with the predictor layer(s) for the
classification}

\item{model}{a trained random forest model as output by
\code{\link{rf_train}}}

\item{classes_file}{filename for predicted classes (or missing)}

\item{prob_file}{filename for predicted probabilities (or missing)}

\item{use_training_flag}{indicates whether to exclude data flagged as
testing data when training the classifier. For this to work the input
train_data \code{data.frame} must have a column named 'training_flag' that
indicates, for each pixel, whether that pixel is a training pixel (coded as
TRUE) or testing pixel (coded as FALSE).}

\item{overwrite}{whether to overwrite \code{out_name} if it already exists}

\item{notify}{notifier to use (defaults to \code{print} function). See the
\code{notifyR} package for one way of sending notifications from R. The
\code{notify} function should accept a string as the only argument.}
}
\value{
a list with 2 elements: the predicted classes as a
\code{RasterLayer} and the class probabilities as a \code{RasterBrick}
}
\description{
Classifies an image using a trained random forest classifier
}
\examples{
#TODO: Fix example
model <- rf_train(x, tr_data)
preds <- rf_classify(x, model)
require(rasterVis)
levelplot(preds$classes)
}

