% Generated by roxygen2 (4.0.2): do not edit by hand
\name{classify}
\alias{classify}
\title{Classify an image using a trained classifier}
\usage{
classify(x, model, classes_file, prob_file, factors = list(),
  overwrite = FALSE)
}
\arguments{
\item{x}{a \code{Raster*} image with the predictor layer(s) for the
classification}

\item{model}{a trained classifier as output by
\code{\link{train_classifier}}}

\item{classes_file}{filename for predicted classes (or missing)}

\item{prob_file}{filename for predicted probabilities (or missing)}

\item{factors}{a list of character vector giving the names of predictors
(layer names from the images used to build \code{train_data}) that should be
treated as factors, and specifying the levels of each factor. For example,
\code{factors=list(year=c(1990, 1995, 2000, 2005, 2010))}.}

\item{overwrite}{whether to overwrite \code{out_name} if it already exists}
}
\value{
a list with 2 elements: the predicted classes as a
\code{RasterLayer} and the class probabilities as a \code{RasterBrick}
}
\description{
This function will produce two outputs - a prediction image and a
probability image. The prediction image contains the predicted classes, the
and the probability image contains the per-pixel predicted probabilities of
occurrence of each class.
}
\details{
This function will run in parallel if a parallel backend is registered with
\code{\link{foreach}}.
}
\examples{
\dontrun{
train_data <- get_pixels(L5TSR_1986, L5TSR_1986_2001_training, "class_1986",
                         training=.6)
model <- train_classifier(train_data)
preds <- classify(L5TSR_1986, model)
plot(preds$classes)
plot(preds$probs)
}
}

