% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/accuracy.R
\docType{class}
\name{error_adj_area-class}
\alias{error_adj_area-class}
\title{A class for error adjusted class areas}
\description{
A class for error adjusted class areas
}
\seealso{
\code{\link{adj_areas}}.
}

