% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pixel_data.R
\docType{methods}
\name{[,pixel_data,character,ANY,ANY-method}
\alias{[,pixel_data,character,ANY,ANY-method}
\title{Extract part of pixel_data class}
\usage{
\S4method{[}{pixel_data,character,ANY,ANY}(x, i, j, ..., drop = TRUE)
}
\arguments{
\item{x}{a \code{pixel_data} object}

\item{i}{a class or list of classes to extract}

\item{j}{unused}

\item{...}{additional arguments (none implemented)}

\item{drop}{unused}
}
\description{
Extract part of pixel_data class
}

