% Generated by roxygen2 (4.0.2): do not edit by hand
\docType{methods}
\name{[,pixel_data,character,ANY,ANY-method}
\alias{[,pixel_data,character,ANY,ANY-method}
\title{Extract part of pixel_data class}
\usage{
\S4method{[}{pixel_data,character,ANY,ANY}(x, i, j, ..., drop = TRUE)
}
\arguments{
\item{x}{a \code{pixel_data} object}

\item{i}{a class or list of classes to extract}

\item{j}{unused}

\item{...}{additional arguments (none implemented)}
}
\description{
Extract part of pixel_data class
}

