\name{get_metadata_item}
\alias{get_metadata_item}
\title{Extract a metadata item from a metadata file in team format}
\usage{
  get_metadata_item(x, item)
}
\arguments{
  \item{x}{a metadata file in team format (ending in .txt)}

  \item{item}{a string giving the name of the metadata item
  to extract}
}
\value{
  The metadata item (as a string)
}
\description{
  Extract a metadata item from a metadata file in team
  format
}

