% Generated by roxygen2 (4.0.2): do not edit by hand
\name{get_metadata_item}
\alias{get_metadata_item}
\title{Extract a metadata item from a metadata file in GDAL PAM format}
\usage{
get_metadata_item(x, key)
}
\arguments{
\item{x}{an image file that has an accompanying GDAL PAM format metadata
file (ending in .aux.xml)}

\item{key}{a string giving the name of the metadata item to extract}
}
\value{
The metadata item (as a string)
}
\description{
GDAL PAM format metadata files end in ".aux.xml".
}

