% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/teamlucc-package.R
\docType{data}
\name{L5TSR_1986}
\alias{L5TSR_1986}
\title{Landsat 5 Surface Reflectance Image from February 6, 1986 (path 15, row 53)}
\description{
Portion of Landsat 5 Surface Reflectance image from the Landsat Climate Data 
Record archive. This subset of the image includes only bands 1-4, and pixel 
values have been scaled by 10000 and rounded off.
}
\seealso{
L5TSR_2001
}

