\docType{package}
\name{LDPKR}
\alias{LDPKR}
\alias{LDPKR-package}
\title{TEAM Remote Sensing Processing Tools}
\description{A set of scripts for processing remote sensing data to support the 
  work of the Tropical Ecology Assessment and Monitoring (TEAM) Network. 
  Intended to be used in conjunction with the set of Python scripts available 
  in the \href{https://github.com/azvoleff/LDPK}{LDPK repository} on github .
}
\author{
Alex Zvoleff <azvoleff@conservation.org>
}
\keyword{package}
