% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/match_rasters.R
\name{match_rasters}
\alias{match_rasters}
\title{Match the coordinate system and extent of two rasters}
\usage{
match_rasters(baseimg, matchimg, filename, method = "bilinear", ...)
}
\arguments{
\item{baseimg}{A /code{Raster*} to use as the base image. This layer will
determine the output coordinate system.}

\item{matchimg}{A /code{Raster*} to match to the base image. If necessary
the /code{matchimg} will be reprojected to match the coordinate system of
the /code{baseimg}. The /code{matchimg} will then be cropped and extended to
match the extent of the /code{baseimg}.}

\item{filename}{file on disk to save \code{Raster*} to (optional)}

\item{method}{the method to use if projection is needed to match image
coordinate systems, or if resampling is needed to align image origins. Can
be "ngb" for nearest-neighbor, or "binlinear" for bilinear interpolation}

\item{...}{additional arguments to pass to \code{writeRaster} (such as
datatype and filename)}
}
\value{
The /code{matchimg} reprojected (if necessary), cropped, and
extended to match the /code{baseimg}.
}
\description{
Match the coordinate system and extent of two rasters
}
\details{
Note that \code{match_rasters} can run in parallel if
\code{beginCluster()} is run prior to running \code{match_rasters}.
}
\examples{
# Mosaic the two ASTER DEM tiles needed to a Landsat image
DEM_mosaic <- mosaic(ASTER_V002_EAST, ASTER_V002_WEST, fun='mean')

# Crop and extend the DEM mosaic to match the Landsat image
matched_DEM <- match_rasters(L5TSR_1986, DEM_mosaic)
}

