% Generated by roxygen2 (4.0.1): do not edit by hand
\name{minnaert_samp}
\alias{minnaert_samp}
\title{Topographic correction for satellite imagery using Minnaert method}
\usage{
minnaert_samp(x, slope, aspect, sunelev, sunazimuth, IL.epsilon = 1e-06,
  slopeclass = NULL, coverclass = NULL, sampleindices = NULL,
  DN_min = NULL, DN_max = NULL)
}
\arguments{
\item{x}{image as a \code{RasterLayer}}

\item{slope}{the slope in radians as a \code{RasterLayer}}

\item{aspect}{the aspect in radians as a \code{RasterLayer}}

\item{sunelev}{sun elevation in degrees}

\item{sunazimuth}{sun azimuth in degrees}

\item{IL.epsilon}{a small amount to add to calculated illumination values
that are equal to zero to avoid division by zero resulting in Inf values}

\item{slopeclass}{the slope classes to calculate k for (in degrees), or
NULL, in which case an algorithm will be used to choose reasonable defaults
for the given image}

\item{coverclass}{used to calculate k for specific cover class (optional)
as \code{RasterLayer}}

\item{sampleindices}{(optional) row-major indices of sample pixels to use in
the calculation of k values for the Minnaert correction. See
\code{\link{gridsample}}.}

\item{DN_min}{minimum allowable pixel value after correction (values less
than \code{DN_min} are set to NA)}

\item{DN_max}{maximum allowable pixel value after correction (values less
than \code{DN_max} are set to NA)}
}
\value{
\code{RasterLayer} with topographically corrected data
}
\description{
Perform topographic correction using the Minnaert method. This code is
modified from the code in the \code{landsat} package written by Sarah
Goslee.  This version of the code has been altered from the \code{landsat}
version to allow the option of using a sample of pixels for calculation of k
in the Minnaert correction (useful when dealing with large images).
}
\details{
See the help page for \code{minnaert} in the \code{landsat} package for
additional details on the parameters.
}
\author{
Sarah Goslee and Alex Zvoleff
}
\references{
Sarah Goslee. Analyzing Remote Sensing Data in {R}: The {landsat} Package.
Journal of Statistical Software, 2011, 43:4, pg 1--25.
http://www.jstatsoft.org/v43/i04/
}

