% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SVIs.R
\docType{methods}
\name{MSAVI2}
\alias{MSAVI2}
\alias{MSAVI2,RasterLayer,RasterLayer,RasterLayer-method}
\alias{MSAVI2,RasterLayer,RasterLayer-method}
\alias{MSAVI2,matrix,matrix,matrix-method}
\alias{MSAVI2,matrix,matrix-method}
\alias{MSAVI2,numeric,numeric,numeric-method}
\alias{MSAVI2,numeric,numeric-method}
\title{Calculates the Modified Soil-Adjusted Vegetation Index (MSAVI)}
\usage{
MSAVI2(red, nir, ...)

\S4method{MSAVI2}{numeric,numeric}(red, nir)

\S4method{MSAVI2}{matrix,matrix}(red, nir)

\S4method{MSAVI2}{RasterLayer,RasterLayer}(red, nir, ...)
}
\arguments{
\item{red}{red}

\item{nir}{near-infrared}

\item{...}{additional arguments as for \code{\link{writeRaster}}}
}
\description{
Note that this avoids the need for calculating L by using the equation for 
MSAVI2 from Qi et al. (1994).
}
\examples{
MSAVI_img <- MSAVI2(red=raster(L5TSR_1986, layer=3),
                    nir=raster(L5TSR_1986, layer=4))
plot(MSAVI_img)
}
\references{
Qi, J., A. Chehbouni, A. R. Huete, Y. H. Kerr, and S.
Sorooshian. 1994. A modified soil adjusted vegetation index. Remote Sensing
of Environment 48:119-126.
}

