% Generated by roxygen2 (4.0.1): do not edit by hand
\name{normalize}
\alias{normalize}
\title{Normalizes two rasters}
\usage{
normalize(x, y, msk, method = "MA", size = ncell(x))
}
\arguments{
\item{x}{a \code{Raster*} to use as the base image}

\item{y}{a \code{Raster*} to normalize to the base image}

\item{msk}{a \code{RasterLayer} with missing values in \code{x} or in {y}
coded as 1, and all other values coded as 0 (optional)}

\item{method}{the regression method to use (must be a method recognized by
\code{lmodel2}}

\item{size}{the number pixels to use in developing the model}
}
\value{
a \code{Raster*} of \code{y} normalized to \code{x}
}
\description{
Performs relative normalization on two rasters using model II regression.
Based on the approach in the \code{relnorm} function in the \code{landsat}
package.
}
\references{
Sarah Goslee. Analyzing Remote Sensing Data in {R}: The {landsat} Package.
Journal of Statistical Software, 2011, 43:4, pg 1--25.
http://www.jstatsoft.org/v43/i04/
}

