% Generated by roxygen2 (4.0.1): do not edit by hand
\docType{class}
\name{pixel_data-class}
\alias{length.pixel_data}
\alias{levels.pixel_data}
\alias{pixel_data-class}
\alias{print.pixel_data}
\alias{print.summary.pixel_data}
\alias{rbind.pixel_data}
\alias{summary.pixel_data}
\title{A class for representing training or testing data}
\usage{
summary.pixel_data(object, ...)

print.summary.pixel_data(x, ...)

\method{length}{pixel_data}(x)

levels.pixel_data(x)

print.pixel_data(x, ...)

rbind.pixel_data(x, ...)
}
\arguments{
\item{object}{a \code{pixel_data} object}

\item{...}{additional arguments (none supported)}

\item{x}{a \code{pixel_data} object}
}
\description{
Used to represent training data for a machine learning classifier for image
classificaion, or testing data used for testing a classification.
}
\section{Slots}{

\describe{
\item{\code{x}}{a \code{data.frame} of independent variables (usually pixel values)}

\item{\code{y}}{a \code{data.frame} of the dependent variable (usually land cover
classes)}

\item{\code{pixel_src}}{a data.framevector used to link pixels in \code{x} and
\code{y} to an input polygon}

\item{\code{training_flag}}{a binary vector of length equal to \code{nrow(x)}
indicating each row in x should be used in training (TRUE) or in testing
(FALSE)}

\item{\code{polys}}{a \code{SpatialPolygonsDataFrame} of the polygons used to choose
the pixels in \code{x} and \code{y}.}
}}

