% Generated by roxygen2 (4.0.0): do not edit by hand
\name{scale_raster}
\alias{scale_raster}
\title{Scales a \code{Raster*} by a power of a given integer and rounds to nearest
integer}
\usage{
scale_raster(x, power_of = 10, max_out = 32767, round_output = TRUE,
  do_scaling = TRUE)
}
\arguments{
\item{x}{a \code{Raster*} object}

\item{power_of}{raster will be scaled using the highest possible power of
this number}

\item{max_out}{the scaling factors will be chosen for each layer to ensure
that the maximum and minimum (if minimum is negative) values of each layer
do not exceed \code{max_out}}

\item{do_scaling}{perform the scaling and return a \code{Raster*} (if
\code{do_scaling} is TRUE) or return a list of scale factors (if
\code{do_scaling} is FALSE)}

\item{round_output}{whether to round the output to the nearest integer}
}
\value{
a \code{Raster*} if \code{do_scaling} is TRUE, or a list of scaling
factors if \code{do_scaling} is false.
}
\description{
Useful for scaling and (optionally) rounding a \code{RasterLayer} to integer
so that a layer can be saved as an integer datatype such as "INT1U",
"INT1S", "INT2" or "INT2S".
}
\seealso{
\code{\link{dataType}}
}

