% Generated by roxygen2 (4.0.1): do not edit by hand
\docType{methods}
\name{src_name}
\alias{src_name}
\alias{src_name,pixel_data-method}
\alias{src_name<-}
\alias{src_name<-,pixel_data-method}
\title{Get src_name for a pixel_data object}
\usage{
src_name(x)

\S4method{src_name}{pixel_data}(x)

src_name(x) <- value

\S4method{src_name}{pixel_data}(x) <- value
}
\arguments{
\item{x}{a \code{pixel_data} object
this pixel_data object}

\item{x}{a \code{pixel_data} object}

\item{src_name}{a new \code{src_name} to assign for pixels in \code{x}}
}
\description{
Get src_name for a pixel_data object

Set src_name for a pixel_data object
}

