% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pixel_data.R
\docType{methods}
\name{src_name}
\alias{src_name}
\alias{src_name,pixel_data-method}
\alias{src_name<-}
\alias{src_name<-,pixel_data-method}
\title{Get or set src_name for a pixel_data object}
\usage{
src_name(x, classes = levels(x@y))

src_name(x) <- value

\S4method{src_name}{pixel_data}(x) <- value
}
\arguments{
\item{x}{a \code{pixel_data} object}

\item{classes}{specifies a subset of classes in \code{x}}

\item{value}{a new \code{src_name} to assign for pixels in \code{x}}
}
\description{
Get or set src_name for a pixel_data object
}

