% Generated by roxygen2 (4.0.0): do not edit by hand
\name{stop_timer}
\alias{stop_timer}
\title{Stop a tracking timer}
\usage{
stop_timer(x, label = "Default")
}
\arguments{
\item{x}{a \code{Track_time} object}

\item{label}{an optional label used to maintain multiple tracking timers}
}
\value{
Track_time object
}
\description{
The \code{label} is optional. If not supplied, the default timer, labelled
'Default' will be used.
}
\examples{
timer <- Track_time()
print(timer)

timer <- start_timer(timer, 'test')

print(timer, 'test')

timer <- stop_timer(timer, 'test')

print(timer)
}
\seealso{
\code{\link{start_timer}}
}

