\name{team_classify}
\alias{team_classify}
\title{Classify a preprocessed surface reflectance image}
\usage{
team_classify(preds)
}
\arguments{
  \item{preds}{the path to an raster stack of predictor
  layers output by the \code{team_preprocess} function}
}
\description{
First the image should be preprocesed using the
\code{team_preprocess} function.
}

