\name{team_fill_gaps}
\alias{team_fill_gaps}
\title{Perform SLC-off gap fill of Landsat 7 ETM+ image}
\usage{
team_fill_gaps()
}
\description{
Calls GNSPI.pro IDL script by Zhu Xiaolin to fill gaps in
SLC-off Landsat 7 ETM+ image.
}
\references{
Zhu, X., Liu, D., Chen, J., 2012. A new geostatistical
approach for filling gaps in Landsat ETM+ SLC-off images.
Remote Sensing of Environment 124, 49-60.
}

