% Generated by roxygen2 (4.0.2): do not edit by hand
\name{tpa2df}
\alias{tpa2df}
\title{Function to convert TIMESAT .tpa binary format file to an R dataframe.}
\usage{
tpa2df(x, max_num_seasons)
}
\arguments{
\item{x}{A string giving the location of a .tpa file output by
TIMESAT}

\item{max_num_seasons}{the maximum number of seasons for any of the pixels
in the file}
}
\value{
A data.frame containing 14 columns: row, col, season, start, end,
length, base_value, peak_time, peak_value, amp, left_deriv, right_deriv,
large_integ, and small_integ
}
\description{
Function to convert TIMESAT .tpa binary format file to an R dataframe.
}
\examples{
# TODO: Need to add examples here, and need to include a sample TIMESAT tpa
# file in the package data.
}

