% Generated by roxygen2 (4.0.1): do not edit by hand
\name{train_classifier}
\alias{train_classifier}
\title{Train a random forest or SVM classifier}
\usage{
train_classifier(train_data, type = "rf", use_training_flag = TRUE,
  train_control = NULL, tune_grid = NULL, use_rfe = FALSE,
  factors = list(), ...)
}
\arguments{
\item{train_data}{a \code{link{pixel_data}} object}

\item{type}{either "svm" (to fit a support vector machine) or "rf" (to fit a
random forest).}

\item{use_training_flag}{indicates whether to exclude data flagged as
testing data when training the classifier. For this to work the input
train_data \code{data.frame} must have a column named 'training_flag' that
indicates, for each pixel, whether that pixel is a training pixel (coded as
TRUE) or testing pixel (coded as FALSE).}

\item{train_control}{default is NULL (reasonable values will be set
automatically).  For details see \code{\link{trainControl}}.}

\item{tune_grid}{the training grid to be used for training the classifier.
See Details.}

\item{use_rfe}{whether to use Recursive Feature Extraction (RFE) as
implemented in the \code{caret} package to select a subset of the input
features to be used in the classification. NOT YET SUPPORTED.}

\item{factors}{a list of character vector giving the names of predictors
(layer names from the images used to build \code{train_data}) that should be
treated as factors, and specifying the levels of each factor. For example,
\code{factors=list(year=c(1990, 1995, 2000, 2005, 2010))}.}

\item{...}{additional arguments (such as \code{ntree} for random forest
classifier) to pass to \code{train}}
}
\value{
a trained model (as a \code{train} object from the \code{caret}
package)
}
\description{
This function trains a Support Vector Machine (SVM) or Random Forest (RF)
classifier for use in an image classification.
}
\details{
For \code{type='svm'}, \code{tunegrid} must be a \code{data.frame} with two
columns: ".sigma" and ".C". For \code{type='rf'}, must be a
\code{data.frame} with one column: '.mtry'.

This function will run in parallel if a parallel backend is registered with
\code{\link{foreach}}.
}
\examples{
train_data <- get_pixels(L5TSR_1986, L5TSR_1986_2001_training, "class_1986",
                         training=.6)
model <- train_classifier(train_data)
}

