% Generated by roxygen2 (4.0.1): do not edit by hand
\name{unstack_ledapscdr}
\alias{unstack_ledapscdr}
\title{Convert Landsat CDR images from HDF4 to ENVI format}
\usage{
unstack_ledapscdr(x, output_folder = NULL, overwrite = FALSE,
  rmhdf = FALSE)
}
\arguments{
\item{x}{input HDF4 file}

\item{output_folder}{output folder (if \code{NULL}, defaults to input folder)}

\item{overwrite}{whether to overwrite existing files}

\item{rmhdf}{whether to remove hdf files after unstacking them}
}
\value{
nothing (used for side effect of converting Landsat CDR HDF files)
}
\description{
This function converts a Landsat surface reflectance (SR) image from the
Landsat Climate Data Record (CDR) archive into a series of single band
images in ENVI format.
}
\details{
Note that this function requires a local GDAL installation. See
http://trac.osgeo.org/gdal/wiki/DownloadingGdalBinaries or
http://trac.osgeo.org/osgeo4w/ to download the appropriate installer for
your operating system.
}
\examples{
\dontrun{
# Unstack files downloaded from CDR:
unstack_ledapscdr('lndsr.LT50150531986021XXX03.hdf')

# Unstack files downloaded from CDR, overwriting any existing files, and
# deleting original HDF files after unstacking:
unstack_ledapscdr('lndsr.LT50150531986021XXX03.hdf', overwrite=TRUE,
                  rmhdf=TRUE)
}
}

